dfac/requirements.txt
pyyaml==6.0.3
requests==2.32.5
typer==0.20.0